+++ src/ui.h
@@ -15,9 +15,9 @@ void ui_HelpMenu(void);
 
 void ui_ScreenWipe(uint8_t);
 
-void ui_CarPicked(uint16_t, uint8_t, uint8_t, gfx_sprite_t **);
+void ui_CarPicked(int, uint8_t, uint8_t, gfx_sprite_t **);
 
-void ui_Cursor(uint16_t, uint16_t, uint8_t);
+void ui_Cursor(int, int, uint8_t);
 
 void ui_VerticalIndicator(int, uint8_t, uint8_t, int);
